added products with Provider
