+++ app/(routes)/category/[slug]/page.tsx
@@ -4,8 +4,12 @@
 import { getCategories as categories } from "@/app/getData/getCategories";
 import { redirect, usePathname } from "next/navigation";
 import { Product } from "@/app/components/Product/Product";
+import { useContext } from "react";
+import { ProductContext } from "@/app/Providers";
 
 const Category = () => {
+  const products = useContext(ProductContext);
+
   const pathName = usePathname();
   const currentPathname = pathName.replace("/category/", "");
 
@@ -17,15 +21,15 @@ const Category = () => {
     redirect("/");
   }
 
-  // const currentProducts = products?.filter(
-  //   (product) => product.slug === currentCategory
-  // );
+  const currentProducts = products?.filter(
+    (product) => product.slug === currentCategory
+  );
 
   return (
     <>
-      {/* {currentProducts?.map((product) => (
+      {currentProducts?.map((product) => (
         <Product key={product.id} {...product} />
-      ))} */}
+      ))}
     </>
   );
 };
