+++ app/(routes)/submit-search/[slug]/page.tsx
@@ -1,20 +1,23 @@
 "use client";
 import { Product } from "@/app/components/Product/Product";
 import { getProductsByInput } from "@/app/getData/getProductsByInput";
+import { ProductContext } from "@/app/Providers";
 import { usePathname } from "next/navigation";
+import { useContext } from "react";
 
 const Page = () => {
+  const products = useContext(ProductContext);
   const pathname = usePathname().replace("/submit-search/", "");
-  const products = getProductsByInput(pathname);
+  const foundProducts = getProductsByInput(pathname, products);
   return (
     <div className="lg:p-10">
-      {products.length === 0 && (
+      {foundProducts.length === 0 && (
         <p>
           there is no products under <b>{pathname}</b>
         </p>
       )}
       <ul className="lg:grid lg:grid-cols-4">
-        {products.map((product) => (
+        {foundProducts.map((product) => (
           <Product key={product.id} {...product} />
         ))}
       </ul>
