+++ app/Providers.tsx
@@ -1,7 +1,35 @@
 "use client";
 import { SessionProvider } from "next-auth/react";
-import { ReactNode } from "react";
+import { ReactNode, useEffect, useState } from "react";
+import { createContext } from "react";
+import { ProductType } from "./types/types";
+export const ProductContext = createContext<ProductType[]>([]);
 
 export const AuthProvider = ({ children }: { children: ReactNode }) => (
   <SessionProvider>{children}</SessionProvider>
 );
+
+export const ProductsProvider = ({ children }: { children: ReactNode }) => {
+  const [products, setProducts] = useState([]);
+  const importProducts = async () => {
+    const res = await fetch("api/getProducts", {
+      method: "GET",
+      headers: {
+        "Content-Type": "application/json",
+      },
+    });
+    if (res.ok) {
+      const { data } = await res.json();
+      setProducts(data);
+    }
+  };
+
+  useEffect(() => {
+    importProducts();
+  }, []);
+  return (
+    <ProductContext.Provider value={products}>
+      {children}
+    </ProductContext.Provider>
+  );
+};
