+++ todo.md
@@ -9,3 +9,5 @@ auth:
 https://www.youtube.com/watch?v=57099pou7iE&list=PLlR2O33QQkfVz-OnrzNTujZD7a-bdo1zY&index=2
 
 add profile that will contain address
+
+add other provider for main layout and pass there products
